split set proc
